refactor: split analysis/subagents/runner.py into focused modules
Extract four focused modules from runner.py (was 256 lines):
- _source_files.py (50 lines): file listing and filtering
- _prompts.py (27 lines): prompt construction
- _pool_launcher.py (71 lines): pool creation and launching
- _evidence_collector.py (62 lines): evidence collection and parsing

runner.py (91 lines) remains as entry point with re-exports for API stability.
